Moved AddItem out of AdminBookings
